Move test to serialization tests
