+++ clients/typescript/test/satellite/serialization.test.ts
@@ -5,6 +5,10 @@ import { Relation, Record } from '../../src/util/types'
 import { DbSchema, TableSchema } from '../../src/client/model/schema'
 import { PgBasicType } from '../../src/client/conversions/types'
 import { HKT } from '../../src/client/util/hkt'
+import Database from 'better-sqlite3'
+import { DatabaseAdapter } from '../../src/drivers/better-sqlite3'
+import { inferRelationsFromSQLite } from '../../src/util/relations'
+import { satelliteDefaults } from '../../src/satellite/config'
 
 test('serialize/deserialize row data', async (t) => {
   const rel: Relation = {
@@ -147,4 +151,70 @@ test('Null mask uses bits as if they were a list', async (t) => {
   const mask = [...s_row.nullsBitmask].map((x) => x.toString(2)).join('')
 
   t.is(mask, '1101000010000000')
+})
+
+test('Prioritize PG types in the schema before inferred SQLite types', async (t) => {
+  const db = new Database(':memory:')
+  t.teardown(() => db.close())
+
+  const adapter = new DatabaseAdapter(db)
+  await adapter.run({
+    sql: 'CREATE TABLE bools (id INTEGER PRIMARY KEY, b INTEGER)',
+  })
+
+  const sqliteInferredRelations = await inferRelationsFromSQLite(
+    adapter,
+    satelliteDefaults
+  )
+  const boolsInferredRelation = sqliteInferredRelations['bools']
+
+  // Inferred types only support SQLite types, so the bool column is INTEGER
+  const boolColumn = boolsInferredRelation.columns[1]
+  t.is(boolColumn.name, 'b')
+  t.is(boolColumn.type, 'INTEGER')
+
+  // Db schema holds the correct Postgres types
+  const boolsDbDescription = new DbSchema(
+    {
+      bools: {
+        fields: new Map([
+          ['id', PgBasicType.PG_INTEGER],
+          ['b', PgBasicType.PG_BOOL],
+        ]),
+        relations: [],
+      } as unknown as TableSchema<
+        any,
+        any,
+        any,
+        any,
+        any,
+        any,
+        any,
+        any,
+        any,
+        HKT
+      >,
+    },
+    []
+  )
+
+  const satOpRow = serializeRow(
+    { id: 5, b: 1 },
+    boolsInferredRelation,
+    boolsDbDescription
+  )
+
+  // Encoded values ["5", "t"]
+  t.deepEqual(satOpRow.values, [
+    new Uint8Array(['5'.charCodeAt(0)]),
+    new Uint8Array(['t'.charCodeAt(0)]),
+  ])
+
+  const deserializedRow = deserializeRow(
+    satOpRow,
+    boolsInferredRelation,
+    boolsDbDescription
+  )
+
+  t.deepEqual(deserializedRow, { id: 5, b: 1 })
 })
